feat : 유저의 정보 삭제를 오직 로그인했고 세션 값이 가진 id와 삭제하고자 하는 id가 일치했을 경우에만 삭제 가능
만약 일치하지 않는 경우 UnauthorizedException 예외 터지고 이를 Advice에서 처리함.

+++ src/main/java/org/zerock/schedule_app_develop/service/UserService.java
@@ -51,7 +51,10 @@ public class UserService {
     }
 
     @Transactional
-    public void deleteById(Long id) {
+    public void deleteById(Long id, LoginSessionAttribute login) {
+        if (login == null || !login.getId().equals(id)) {
+            throw new UnauthorizedException("Unauthorized");
+        }
         userRepository.deleteById(id);
     }
